+++ python/lsst/ts/m2com/mock/__init__.py
@@ -25,5 +25,6 @@ from .mock_command import *
 from .mock_script_engine import *
 from .mock_control_open_loop import *
 from .mock_control_closed_loop import *
+from .mock_error_handler import *
 from .mock_model import *
 from .mock_server import *